Fix bug concerning one block non-standardized models and version update
This commit fixes a bug that led to crashes when using models without
standardization with only one X block, because the block was not wrapped
in a list. In addition, the package requirement of scikit-learn is
updated to 0.20.0, because previous releases do not support
standardization of matrices with missing values.
	modified:   mbpls/__init__.py
	modified:   mbpls/mbpls.py
	modified:   setup.py