Change the behavior of the transform method and add Scikit-learn estimator check
The standard behavior of the transform method was changed to only return
the superscores and response scores. The block scores are only returned
when the return_block_scores parameter of transform is set to true.
Furthermore, the Scikit-learn estimator check was added to the test
script, to test the compatibility of the mbpls module with Scikit-learn
methods.
Fixes #6
	modified:   mbpls/mbpls.py
	modified:   mbpls/tests/test_mbpls.py